+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes
 * [UI]: Fixed bug where SVG files could not be exported through the advanced visualization page. (0.21.2)
 * [UI]: Fixed bug where users could not share more than nine samples. (0.21.2)
 * [UI]: Fixed bug where edit groups page would throw a server exception. (0.21.3)
+* [Workflow]: Fixed bug where auto updating metadata from analysis submission failed for non-admin user. (0.21.4)
 
 0.20.0 to 0.21.0
 ----------------